requirements.txt
psycopg2-binary
spotipy
json